src/airflow/requirements.txt
pandas==2.2.2
pyarrow==16.1.0
psycopg2-binary==2.9.9

